server/pyproject.toml
[tool.poetry]
name = "document-summariser-server"
version = "0.1.0"
description = ""
authors = ["Michael B <michael@ramyun.co.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
openai = {extras = ["embeddings"], version = "^0.27.4"}
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"
nltk = "^3.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
